Create mean_median_mode.r
Inspiration taken from hackerrank. Modified version due to the mode on hackerrank not being the true mode. This version will, if there is no apparent mode, return "No Mode" (ie when all the values in a data set occur the same time). In addition it will average the values that occur the most and the same amount of times.